requirements.txt
pandas==1.1.4
dlib==19.21.1
streamlit==0.82.0
numpy==1.19.4
matplotlib==3.3.1
playsound==1.2.2
scipy==1.5.3
imutils==0.5.3
pyserial==3.4
scikit-learn==0.23.2
wheel==0.35.1
yaml==0.2.5
tornado==6.1
seaborn==0.11.0
libblas==3.8.0